asset downloader and updater rewritten with asyncio

+++ subsync/assets/updater.py
@@ -1,5 +1,6 @@
 import config
 import utils
+import error
 import json
 import shutil
 import os
@@ -10,52 +11,44 @@ import logging
 logger = logging.getLogger(__name__)
 
 
-class Updater(object):
-    def __init__(self):
-        self.upgradeReady = False
-        self.install = None
-        self.version = None
-        self.load()
-
-    def load(self):
+def getLocalUpdate():
+    try:
         dirPath = os.path.join(config.assetdir, 'upgrade')
-        path = os.path.join(dirPath, 'upgrade.json')
-
         if os.path.isdir(dirPath):
-            try:
-                with open(path, encoding='utf8') as fp:
-                    upgrade = json.load(fp)
+            path = os.path.join(dirPath, 'upgrade.json')
+            with open(path, encoding='utf8') as fp:
+                update = json.load(fp)
 
-                self.install = os.path.join(dirPath, upgrade['install'])
-                self.version = utils.parvseVersion(upgrade['version'])
+            return dict(
+                    version = utils.parseVersion(update['version']),
+                    path = os.path.join(dirPath, update['install']),
+                    cwd = dirPath)
 
-                currentVersion = utils.getCurrentVersion()
-                if currentVersion and self.version > currentVersion:
-                    self.upgradeReady = True
-                else:
-                    self.upgradeReady = False
-                    self.remove()
+    except Exception as e:
+        logger.warning('read update info failed, %r', e, exc_info=True)
+        removeLocalUpdate()
 
-            except Exception as e:
-                logger.error('invalid upgrade description, %r', e)
-                self.remove()
-
-    def remove(self):
-        self.upgradeReady = False
+def removeLocalUpdate():
+    try:
         path = os.path.join(config.assetdir, 'upgrade')
-        try:
+        if os.path.isdir(path):
+            logger.info('removing update from %s', path)
             shutil.rmtree(path, ignore_errors=True)
-        except Exception as e:
-            logger.error('cannot remove upgrade data from %s: %r', path, e)
 
-    def upgrade(self):
-        cwd = os.path.join(config.assetdir, 'upgrade')
-        path = os.path.join(cwd, self.install)
+    except Exception as e:
+        logger.error('cannot remove upgrade data from %s: %r', path, e)
+
+def installLocalUpdate():
+    try:
+        update = getLocalUpdate()
+        path = update['path']
+
         logger.info('executing installer %s', path)
-
         mode = os.stat(path).st_mode
         if (mode & stat.S_IEXEC) == 0:
             os.chmod(path, mode | stat.S_IEXEC)
-        subprocess.Popen(path, cwd=cwd)
-        self.upgradeReady = False
+        subprocess.Popen(path, cwd=update['cwd'])
 
+    except Exception as e:
+        logger.error('cannot install update %s: %r', path, e)
+        raise error.Error(_('Update instalation failed miserably'))
